CC2_V16 - Generation of original fragment counts file - Logging updates - ICustomFragmentCount now uses an int like the original EntryData

+++ CustomCraftSML/Interfaces/ICustomFragmentCount.cs
@@ -2,6 +2,6 @@
 {
     interface ICustomFragmentCount : ITechTyped
     {
-        short FragmentsToScan { get; }
+        int FragmentsToScan { get; }
     }
 }
